add the other Tor tests into tests_gui and refactor them. Reinstate the shell script for the sake of Travis
