+++ tests_gui/run_unit_tests.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+for test in `ls -1 | egrep ^local_`; do
+  pytest $test -vvv || exit 1
+done
